Simply the awk code and add a safeguard to xargs
Previously, "go get" would return an error when the project has no
dependencies. Adding the "-r" (or "--no-run-if-empty") option to xargs
fixes this by not invoking the "go get" command when the input is empty.
The option is a GNU extension, thus not portable. But, as Docker
typically runs on Linux, this may not be a problem.